Use an image tag that matches the dagster version and python version packaged in the deps.pex (#53)
The builder points the code location to a docker image tag and a `pex_tag`. This change ensures that the dagster version installed in the docker image is identical to the dagster version shipped in the deps.pex. This change also selects an image matching the python version used in user code.