Add a unique suffix to each built image tag (#110)
add a unique suffix to each built image tag

Summary:
Previously we assumed that the only time the image tag needed to change was when your code changed / the github SHA of your main github repository changed - this is not true in the case where you are pulling in code from some other repository. This can cause confusing caching issues, particularly if multiple runs with the same image tag can exist at the same time.

This PR changes things so that by default we now append a unique 8-character suffix to each image tag.