Increase location-load-timeout in cloud action
Summary:
This is currently set to 10 minutes, which is typically plenty - but can be exceeded in certain situations (ECS plus a slow code location load). It's also not clear whether this particular timeout is giving us any benefit in practice - the agent will almost always either be a) down, in which case the heartbeat timeout will kick in or b) will have its own timeout on the agent side that will move the location into an error state. So there's not much downside in increasing this.

Test Plan:
Rebuild and point an action at the new image, verify that when there's an error the github action still errors without timing out.