Add base_image support to the serverless_prod_deploy action
Summary:
Allows you to set build:base_image to vary the base image per code location.

Would like to port this to the new "dagster-cloud ci" commands as well.

Test Plan:
Build a release branch and test a build with and without the base_image set, in both a dagster_cloud.yaml file and in the action itself.