Use correct command to upgrade pip in dbt nux
Summary:
The previous command did not actually upgrade pip and left us on 20.0.2.

Test Plan:
Go through the NUX with this new YAML, it now correctly backtracks and is able to install the correct set of dependencies.

I'm unclear if I need to do a release of the action for this to be applied or not?