No longer build the github action using python 3.8 (#213)
Summary:
This is needed to be able to deploy newer versions of the dagster and dagster-cloud CLI.

Once we do this, we can no longer push to v0.1 or we may break users on python 3.8 / older versions. We need to move this action to a weekly release schedule anyway, so this is the opportunity to do that.