deploy an arm64 github action image as well as an amd64 image
summary:
The PEX is already set up to work on both amd64 and arm64, but users running our image (these days i think this is mostly just Gitlab users) on arm64 aren't able to use the built image.

We could also do multi-platform builds but that seems non-trivial to set up in our CI, so starting with separate tags for now.

stop using a PEX here